:truck: Rename db-client-legacy to db-migration-client.
Why? To better reflect its purpose.
