+++ api/src/routes/migrate-router.ts
@@ -1,6 +1,6 @@
 import { Router, type Request, type Response } from "express"
 
-import dbLegacy from "@/db/db-client-legacy"
+import dbLegacy from "@/db/db-migration-client"
 import logger from "@/utils/logger"
 
 export const migrateRouter = Router()
